Add support to auto despawn amount in frames
Currently, it was only possible to specify in when an entity should be removed
in seconds as a floating point. It can also be important to specify this duration
in frames in case the programmer wants to delete an entity in two frames in to
keep it for a very short period of time.